fixed some compatibility issues introduced with merge. Need to get float16 working with diffmotions

+++ python_visual_mpc/video_prediction/prediction_model_sawyer.py
@@ -392,7 +392,8 @@ class Prediction_Model(object):
 
                 if 'visual_flowvec' in self.conf:
                     motion_vecs = self.compute_motion_vector(cdna_kerns)
-                    output = tf.zeros([self.conf['batch_size'],64,64,2])
+
+                    output = tf.zeros([self.conf['batch_size'],64,64,2], dtype = motion_vecs[0].dtype)
                     for vec, mask in zip(motion_vecs, mask_list[1:]):
                         vec = tf.reshape(vec, [self.conf['batch_size'], 1, 1, 2])
                         vec = tf.tile(vec, [1, 64,64, 1])
@@ -506,8 +507,8 @@ class Prediction_Model(object):
         dc = np.expand_dims(dc, axis=0)
         dc = np.repeat(dc, self.conf['kern_size'], axis=0)
         dr = np.transpose(dc)
-        dr = tf.constant(dr, dtype=tf.float32)
-        dc = tf.constant(dc, dtype=tf.float32)
+        dr = tf.constant(dr, dtype=cdna_kerns.dtype)
+        dc = tf.constant(dc, dtype=cdna_kerns.dtype)
 
         cdna_kerns = tf.transpose(cdna_kerns, [2, 3, 0, 1])
         cdna_kerns = tf.split(cdna_kerns, self.conf['num_masks'], axis=1)
